+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortalUserController.test.js
@@ -158,7 +158,7 @@ describe('PortalUserController', () => {
                 accept: 'text/*',
             },
             query: {
-                redirect: '/foo/bar',
+                redirectUrl: '/foo/bar',
             },
             pluginContext: {
                 loggerFactory: dummyLoggerFactory,