+++ packages/plugin-packages/mashroom-security-provider-simple/test/MashroomSimpleSecurityProvider.test.js
@@ -27,7 +27,7 @@ describe('MashroomSimpleSecurityProvider', () => {
 
         expect(result).toBeTruthy();
         expect(result.status).toBe('deferred');
-        expect(redirectUrl).toBe('/login?ref=L2Zvby9iYXI=');
+        expect(redirectUrl).toBe('/login?redirectUrl=%2Ffoo%2Fbar');
     });
 
     it('passes the authentication hints to the login page', async () => {
@@ -52,7 +52,7 @@ describe('MashroomSimpleSecurityProvider', () => {
 
         expect(result).toBeTruthy();
         expect(result.status).toBe('deferred');
-        expect(redirectUrl).toBe('/login?ref=L2Zvby9iYXI=&hint1=foo&hint2=2');
+        expect(redirectUrl).toBe('/login?redirectUrl=%2Ffoo%2Fbar&hint1=foo&hint2=2');
     });
 
     it('processes the form login correctly', async () => {